map_object_spawner
